// TODO : passer index_table en int plutot que long : done
git-svn-id: https://svn.ias.u-psud.fr/sanepic/trunk@302 339b86bc-bb6a-0410-a225-86439261ec80

+++ saneFrameOrder/src/main.cpp
@@ -64,7 +64,7 @@ int main(int argc, char *argv[])
 
 		std::vector<string> fitsvect;
 		std::vector<string> noisevect;
-		std::vector<long> scans_index;
+		std::vector<int> scans_index;
 		//	vector<long>::iterator it;
 
 		//std::vector<long> fframes_vec, nsamples_vec;
